backend-services/auth_service/requirements.txt
django>=5.2,<6.0
djangorestframework
djangorestframework-simplejwt
django-cors-headers
pymysql
cryptography
python-dotenv
uvicorn[standard]
